refactor: relocate solutions into strategy based packages
- Updated package declarations and preserved test integrity
- This reorganization improves solution discoverability and aligns with a scalable, strategy-based file structure
